backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
supabase==2.4.6
httpx==0.27.0
python-jose==3.5.0
python-dotenv==1.0.1
aiofiles==23.2.1
